Move usersService
